Add option for a dedicated HDMI_CEC_RX pin.
On ESP32 (and its variants), if a pin is switched from INPUT/INPUT_PULLUP
to OUTPUT, its attached interrupt stops firing.

As a workaround, we can connect 2 pins together and use one for TX(setLineState)
and one for RX (attachInterruptArg).

If a dedicated RX pin is not specified in the module configuration, the
same pin is used for TX and RX. (ESP8266 friendly.)

Resolves https://github.com/arendst/Tasmota/discussions/21874

Additionally, the IRAM_ATTR should be specified only in the function
declaration, not at the function definition. This commit fixes that.